test(core): split build result module code tests
